Register `before_request` and `after_request` globally
The blueprint is a global object.

Flask no longer wants you to mutate the global object by repeatedly
calling `before_request` or `after_request` on it inside the
`create_app` function (`create_app` gets called every time we run a test
I think, but the blueprint itself is shared between test runs).

This moves all that code to a place where it’s only run once.
